Create yearly_timeline.md
This creates a yearly timeline document so we can keep track of things we have to do every year as a team and make sure none of it falls through the cracks. It's just a bullet list, and the intention is that each one of the items will become a link as we fill out more playbooks for each one.